lesson editor added
